+++ src/TOML/Class.hs
@@ -3,20 +3,36 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module TOML.Class
-  ( TValue (..),
+  ( Value (..),
   )
 where
 
-import qualified Data.Time as T
-import Pretty (Pretty (..))
+import Data.Scientific (Scientific)
+import Data.Text (Text)
+import Data.Time (Day, LocalTime, TimeOfDay, ZonedTime)
+import qualified Data.Time as Time
 
--- |
--- * Date and times follow RFC 3339.
-data TValue
-  = ZonedTime T.ZonedTime
+data Value
+  = Bool Bool
+  | Integer Int
+  | Double Scientific
+  | Text Text
+  | ZonedTime ZonedTime
+  | LocalTime LocalTime
+  | Day Day
+  | TimeOfDay TimeOfDay
+  | Array [Value]
+  deriving stock (Show)
 
-instance Show TValue where
-  show = undefined
-
-instance Pretty TValue where
-  pretty = undefined
+instance Eq Value where
+  (Bool b1) == (Bool b2) = b1 == b2
+  (Integer i1) == (Integer i2) = i1 == i2
+  (Double f1) == (Double f2) = f1 == f2
+  (Text s1) == (Text s2) = s1 == s2
+  (ZonedTime a) == (ZonedTime b) =
+    Time.zonedTimeToUTC a == Time.zonedTimeToUTC b
+  (LocalTime a) == (LocalTime b) = a == b
+  (Day a) == (Day b) = a == b
+  (TimeOfDay a) == (TimeOfDay b) = a == b
+  (Array a1) == (Array a2) = a1 == a2
+  _ == _ = False
